Migrate chat-rooms.js to TypeScript with type definitions and improvements
